rag-api/src/api/requirements.txt
python-dotenv
langchain_groq
langchain
fastapi
uvicorn